refactor(Bsale::Spider): add parent class
